Move API server to its own package
