+++ go.mod
@@ -6,4 +6,5 @@ require (
 	github.com/golang/glog v0.0.0-20210429001901-424d2337a529
 	github.com/rabbitmq/amqp091-go v0.0.0-20210714180937-de74e8a7d0e0
 	github.com/rabbitmq/rabbitmq-stream-go-client v0.0.0-20210720070408-b04b01a752c3
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
 )
